Update [id].js
add authentication for updating and deleting an order

+++ pages/api/products/[id].js
@@ -7,9 +7,12 @@ export default async function handler(req, res) {
 
   const { 
     method,
-    query: { id } 
+    query: { id },
+    cookies,
   } = req
 
+  const token = cookies.token
+  
   // GET Method for getting a product
   if(method === "GET"){
     try {
@@ -22,7 +25,14 @@ export default async function handler(req, res) {
 
   // PUT Method for updating product
   if(method === "PUT"){
+    if(!token || token !== process.env.TOKEN){
+      return res.status(401).json("Not Authenticated!")
+    }
     try{
+      const product = await Product.findByIdAndUpdate(id, req.body, {
+        new: true
+      })
+      res.status(201).json(product)
     }catch(err){
       res.status(500).json(err)
     }
@@ -30,6 +40,9 @@ export default async function handler(req, res) {
 
   // DELETE Method for deleting product
   if(method === "DELETE"){
+    if(!token || token !== process.env.TOKEN){
+      return res.status(401).json("Not Authenticated!")
+    }
     try{
       await Product.findByIdAndDelete(id)
       res.status(200).json("The product has been deleted successfully!")
